[TS-338] Refactor: rest record page refactoring
